+++ src/tests/EntityGraphQL.EF.Tests/TestDbContext.cs
@@ -29,6 +29,7 @@ public class Actor(string name) : IEntityWithId
     public string Name { get; set; } = name;
     public DateTime Birthday { get; set; }
     public List<Movie> Movies { get; set; } = [];
+    public List<Actor> Children { get; set; } = [];
 }
 
 public class Director(string name) : IEntityWithId